publish-artifacts: move s3 logic to pkg/release
Previously, the S3 logic was in both publish-artifacts and
publish-provisional-artifacts. Additionally there was some logic in
pkg/release.

This PR refators these 3 modules:
* Move logic to separate files.
* No direct s3 imports from publish-artifacts and
  publish-provisional-artifacts.
* Create an extra abstraction interface to hide the s3-specific structs
  and calls.
* The "latest" URLs in publish-provisional-artifacts use S3's website
  redirect instead of copying the whole binary. This unifies it with
  publish-artifacts.

Release note: None
