+++ pkg/cmd/publish-provisional-artifacts/main_test.go
@@ -22,10 +22,8 @@ import (
 	"unicode/utf8"
 
 	"github.com/alessio/shellescape"
-	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/cockroachdb/cockroach/pkg/release"
 	"github.com/cockroachdb/cockroach/pkg/testutils"
-	"github.com/cockroachdb/errors"
 	"github.com/stretchr/testify/require"
 )
 
@@ -34,41 +32,41 @@ type mockS3 struct {
 	puts []string
 }
 
-var _ s3I = (*mockS3)(nil)
+var _ release.ObjectPutGetter = (*mockS3)(nil)
 
-func (s *mockS3) GetObject(i *s3.GetObjectInput) (*s3.GetObjectOutput, error) {
+func (s *mockS3) GetObject(i *release.GetObjectInput) (*release.GetObjectOutput, error) {
 	url := fmt.Sprintf(`s3://%s/%s`, *i.Bucket, *i.Key)
 	s.gets = append(s.gets, url)
-	o := &s3.GetObjectOutput{
+	o := &release.GetObjectOutput{
 		Body: ioutil.NopCloser(bytes.NewBufferString(url)),
 	}
 	return o, nil
 }
 
-func (s *mockS3) PutObject(i *s3.PutObjectInput) (*s3.PutObjectOutput, error) {
+func (s *mockS3) PutObject(i *release.PutObjectInput) error {
 	url := fmt.Sprintf(`s3://%s/%s`, *i.Bucket, *i.Key)
 	if i.CacheControl != nil {
 		url += `/` + *i.CacheControl
 	}
 	if i.Body != nil {
-		bytes, err := ioutil.ReadAll(i.Body)
+		binary, err := ioutil.ReadAll(i.Body)
 		if err != nil {
-			return nil, err
+			return err
 		}
 		if strings.HasSuffix(*i.Key, release.ChecksumSuffix) {
 			// Unfortunately the archive tarball checksum changes every time,
 			// because we generate tarballs and the copy file modification time from the generated files.
 			// This makes the checksum not reproducible.
 			s.puts = append(s.puts, fmt.Sprintf("%s CONTENTS <sha256sum>", url))
-		} else if utf8.Valid(bytes) {
-			s.puts = append(s.puts, fmt.Sprintf("%s CONTENTS %s", url, bytes))
+		} else if utf8.Valid(binary) {
+			s.puts = append(s.puts, fmt.Sprintf("%s CONTENTS %s", url, binary))
 		} else {
 			s.puts = append(s.puts, fmt.Sprintf("%s CONTENTS <binary stuff>", url))
 		}
 	} else if i.WebsiteRedirectLocation != nil {
 		s.puts = append(s.puts, fmt.Sprintf("%s REDIRECT %s", url, *i.WebsiteRedirectLocation))
 	}
-	return &s3.PutObjectOutput{}, nil
+	return nil
 }
 
 type mockExecRunner struct {
@@ -80,8 +78,8 @@ func (r *mockExecRunner) run(c *exec.Cmd) ([]byte, error) {
 	if r.fakeBazelBin == "" {
 		panic("r.fakeBazelBin not set")
 	}
-	if c.Dir == `` {
-		return nil, errors.Errorf(`Dir must be specified`)
+	if c.Dir == "" {
+		return nil, fmt.Errorf("`Dir` must be specified")
 	}
 	cmd := fmt.Sprintf("env=%s args=%s", c.Env, shellescape.QuoteCommand(c.Args))
 	r.cmds = append(r.cmds, cmd)
@@ -286,15 +284,15 @@ func TestProvisional(t *testing.T) {
 			defer cleanup()
 
 			var s3 mockS3
-			var exec mockExecRunner
+			var runner mockExecRunner
 			fakeBazelBin, cleanup := testutils.TempDir(t)
 			defer cleanup()
-			exec.fakeBazelBin = fakeBazelBin
+			runner.fakeBazelBin = fakeBazelBin
 			flags := test.flags
 			flags.pkgDir = dir
-			execFn := release.ExecFn{MockExecFn: exec.run}
-			run(&s3, flags, execFn)
-			require.Equal(t, test.expectedCmds, exec.cmds)
+			execFn := release.ExecFn{MockExecFn: runner.run}
+			run([]release.ObjectPutGetter{&s3}, flags, execFn)
+			require.Equal(t, test.expectedCmds, runner.cmds)
 			require.Equal(t, test.expectedGets, s3.gets)
 			require.Equal(t, test.expectedPuts, s3.puts)
 		})
@@ -325,24 +323,20 @@ func TestBless(t *testing.T) {
 				isRelease: true,
 				branch:    `provisional_201901010101_v0.0.1`,
 			},
-			expectedGets: []string{
-				"s3://binaries.cockroachdb.com/cockroach-v0.0.1.linux-amd64.tgz",
-				"s3://binaries.cockroachdb.com/cockroach-v0.0.1.linux-amd64.tgz.sha256sum",
-				"s3://binaries.cockroachdb.com/cockroach-v0.0.1.darwin-10.9-amd64.tgz",
-				"s3://binaries.cockroachdb.com/cockroach-v0.0.1.darwin-10.9-amd64.tgz.sha256sum",
-				"s3://binaries.cockroachdb.com/cockroach-v0.0.1.windows-6.2-amd64.zip",
-				"s3://binaries.cockroachdb.com/cockroach-v0.0.1.windows-6.2-amd64.zip.sha256sum",
-			},
+			expectedGets: nil,
 			expectedPuts: []string{
 				"s3://binaries.cockroachdb.com/cockroach-latest.linux-amd64.tgz/no-cache " +
-					"CONTENTS s3://binaries.cockroachdb.com/cockroach-v0.0.1.linux-amd64.tgz",
-				"s3://binaries.cockroachdb.com/cockroach-latest.linux-amd64.tgz.sha256sum/no-cache CONTENTS <sha256sum>",
+					"REDIRECT cockroach-v0.0.1.linux-amd64.tgz",
+				"s3://binaries.cockroachdb.com/cockroach-latest.linux-amd64.tgz.sha256sum/no-cache " +
+					"REDIRECT cockroach-v0.0.1.linux-amd64.tgz.sha256sum",
 				"s3://binaries.cockroachdb.com/cockroach-latest.darwin-10.9-amd64.tgz/no-cache " +
-					"CONTENTS s3://binaries.cockroachdb.com/cockroach-v0.0.1.darwin-10.9-amd64.tgz",
-				"s3://binaries.cockroachdb.com/cockroach-latest.darwin-10.9-amd64.tgz.sha256sum/no-cache CONTENTS <sha256sum>",
+					"REDIRECT cockroach-v0.0.1.darwin-10.9-amd64.tgz",
+				"s3://binaries.cockroachdb.com/cockroach-latest.darwin-10.9-amd64.tgz.sha256sum/no-cache " +
+					"REDIRECT cockroach-v0.0.1.darwin-10.9-amd64.tgz.sha256sum",
 				"s3://binaries.cockroachdb.com/cockroach-latest.windows-6.2-amd64.zip/no-cache " +
-					"CONTENTS s3://binaries.cockroachdb.com/cockroach-v0.0.1.windows-6.2-amd64.zip",
-				"s3://binaries.cockroachdb.com/cockroach-latest.windows-6.2-amd64.zip.sha256sum/no-cache CONTENTS <sha256sum>",
+					"REDIRECT cockroach-v0.0.1.windows-6.2-amd64.zip",
+				"s3://binaries.cockroachdb.com/cockroach-latest.windows-6.2-amd64.zip.sha256sum/no-cache " +
+					"REDIRECT cockroach-v0.0.1.windows-6.2-amd64.zip.sha256sum",
 			},
 		},
 	}
@@ -351,7 +345,7 @@ func TestBless(t *testing.T) {
 		t.Run(test.name, func(t *testing.T) {
 			var s3 mockS3
 			var execFn release.ExecFn // bless shouldn't exec anything
-			run(&s3, test.flags, execFn)
+			run([]release.ObjectPutGetter{&s3}, test.flags, execFn)
 			require.Equal(t, test.expectedGets, s3.gets)
 			require.Equal(t, test.expectedPuts, s3.puts)
 		})
