+++ pkg/release/BUILD.bazel
@@ -2,11 +2,18 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "release",
-    srcs = ["release.go"],
+    srcs = [
+        "build.go",
+        "release.go",
+        "s3.go",
+        "upload.go",
+    ],
     importpath = "github.com/cockroachdb/cockroach/pkg/release",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/build/util",
+        "@com_github_aws_aws_sdk_go//aws",
+        "@com_github_aws_aws_sdk_go//aws/session",
         "@com_github_aws_aws_sdk_go//service/s3",
         "@com_github_cockroachdb_errors//:errors",
     ],
